+++ src/komm/_error_control_block/CyclicCode.py
@@ -43,13 +43,13 @@ class CyclicCode(BlockCode):
         >>> (code.length, code.dimension, code.redundancy)
         (23, 12, 11)
         >>> code.minimum_distance
-        7
+        np.int64(7)
 
         >>> code = komm.CyclicCode(length=23, check_polynomial=0b1010010011111)  # Golay (23, 12)
         >>> (code.length, code.dimension, code.redundancy)
         (23, 12, 11)
         >>> code.minimum_distance
-        7
+        np.int64(7)
     """
     _length: int = field(default=None, repr=False, alias="length")
     _generator_polynomial: BinaryPolynomial | int = field(
