+++ src/komm/_error_control_block/TerminatedConvolutionalCode.py
@@ -39,7 +39,7 @@ class TerminatedConvolutionalCode(BlockCode):
 
         >>> code = komm.TerminatedConvolutionalCode(convolutional_code, num_blocks=3, mode='zero-termination')
         >>> (code.length, code.dimension, code.minimum_distance)
-        (8, 3, 3)
+        (np.int64(8), 3, np.int64(3))
         >>> code.generator_matrix
         array([[1, 1, 0, 1, 0, 0, 0, 0],
                [0, 0, 1, 1, 0, 1, 0, 0],
@@ -47,7 +47,7 @@ class TerminatedConvolutionalCode(BlockCode):
 
         >>> code = komm.TerminatedConvolutionalCode(convolutional_code, num_blocks=3, mode='direct-truncation')
         >>> (code.length, code.dimension, code.minimum_distance)
-        (6, 3, 2)
+        (6, 3, np.int64(2))
         >>> code.generator_matrix
         array([[1, 1, 0, 1, 0, 0],
                [0, 0, 1, 1, 0, 1],
@@ -55,7 +55,7 @@ class TerminatedConvolutionalCode(BlockCode):
 
         >>> code = komm.TerminatedConvolutionalCode(convolutional_code, num_blocks=3, mode='tail-biting')
         >>> (code.length, code.dimension, code.minimum_distance)
-        (6, 3, 3)
+        (6, 3, np.int64(3))
         >>> code.generator_matrix
         array([[1, 1, 0, 1, 0, 0],
                [0, 0, 1, 1, 0, 1],
